+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensions.kt
@@ -5,27 +5,46 @@ package com.windea.breezeframework.core.extensions
 
 import com.windea.breezeframework.core.annotations.api.*
 import com.windea.breezeframework.core.enums.core.*
+import java.util.*
 import java.util.concurrent.*
-import kotlin.random.*
+import kotlin.collections.ArrayList
+import kotlin.collections.HashMap
+import kotlin.random.Random
 
 //region portal extensions
-/**构建线程安全的并发列表。*/
+/**构建一个空的枚举集。*/
+inline fun <reified T : Enum<T>> enumSetOf(): EnumSet<T> = EnumSet.noneOf(T::class.java)
+
+/**构建一个枚举集。*/
+fun <T : Enum<T>> enumSetOf(first: T, vararg elements: T): EnumSet<T> = EnumSet.of(first, *elements)
+
+/**构建一个包含所有枚举值的枚举集。*/
+inline fun <reified T : Enum<T>> enumSetOfAll(): EnumSet<T> = EnumSet.allOf(T::class.java)
+
+/**构建一个空的枚举映射。*/
+inline fun <reified K : Enum<K>, V> enumMapOf(): EnumMap<K, V> = EnumMap(K::class.java)
+
+/**构建一个枚举映射。*/
+fun <K : Enum<K>, V> enumMapOf(vararg pairs: Pair<K, V>): EnumMap<K, V> = EnumMap(pairs.toMap())
+
+
+/**构建一个空的线程安全的并发列表。*/
 inline fun <T> concurrentListOf(): CopyOnWriteArrayList<T> = CopyOnWriteArrayList()
 
-/**构建线程安全的并发列表。*/
-inline fun <T> concurrentListOf(vararg elements: T): CopyOnWriteArrayList<T> = CopyOnWriteArrayList(elements)
+/**构建一个线程安全的并发列表。*/
+fun <T> concurrentListOf(vararg elements: T): CopyOnWriteArrayList<T> = CopyOnWriteArrayList(elements)
 
-/**构建线程安全的并发集。*/
+/**构建一个空的线程安全的并发集。*/
 inline fun <T> concurrentSetOf(): CopyOnWriteArraySet<T> = CopyOnWriteArraySet()
 
-/**构建线程安全的并发集。*/
-inline fun <T> concurrentSetOf(vararg elements: T): CopyOnWriteArraySet<T> = CopyOnWriteArraySet(elements.toSet())
+/**构建一个线程安全的并发集。*/
+fun <T> concurrentSetOf(vararg elements: T): CopyOnWriteArraySet<T> = CopyOnWriteArraySet(elements.toSet())
 
-/**构建线程安全的并发映射。*/
+/**构建一个空的线程安全的并发映射。*/
 inline fun <K, V> concurrentMapOf(): ConcurrentMap<K, V> = ConcurrentHashMap()
 
-/**构建线程安全的并发映射。*/
-inline fun <K, V> concurrentMapOf(vararg pairs: Pair<K, V>): ConcurrentMap<K, V> = ConcurrentHashMap(pairs.toMap())
+/**构建一个线程安全的并发映射。*/
+fun <K, V> concurrentMapOf(vararg pairs: Pair<K, V>): ConcurrentMap<K, V> = ConcurrentHashMap(pairs.toMap())
 //endregion
 
 //region operator overrides